+++ Mul-Ty-Player Client/Classes/GamePlay/HardcoreMode/HardcoreModeHandler.cs
@@ -30,6 +30,7 @@ public class HardcoreModeHandler
     
     public void InitializeLevel(int level)
     {
+        HDC_IcicleHandler.ResetIcicles();
         SyncHandler.HThEg.ShowStopwatch();
         var triggerSphereCount = Client.HSync.HTrigger.GetTriggerSphereCount();
         for (var triggerIndex = 0; triggerIndex < triggerSphereCount; triggerIndex++)
@@ -149,7 +150,7 @@ public class HardcoreModeHandler
             ProcessHandler.TryRead(0x28CB58, out int taTime, true, "TA Time");
             ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x28CB58, BitConverter.GetBytes((int)(taTime * 0.7)));
         }
-        
+
         Client.HHero.SetRunSpeed(12.5f);
         Client.HHero.SetAirSpeed(12.5f);
         Client.HHero.SetSwimSpeed(25f);
@@ -159,9 +160,11 @@ public class HardcoreModeHandler
         Client.HHero.SetFallDelta(650f);
         Client.HHero.SetOpalMagnetisation(false);
         Client.HHardcore.SetEnemySpeedMultiplier(2.0f);
-        Client.HHero.SetWaterSlideSpeeds(15f, 15f, 30f);
+        Client.HHero.SetWaterSlideSpeeds(20f, 20f, 30f);
         if (Client.HLevel.CurrentLevelId != Levels.OutbackSafari.Id)
             AttributeHandler.SetBoomerangRange(1200f);
+        if (Client.HHero.IsBull())
+            Client.HHero.SetBushpigSpeeds(50f, 0.6f, 1.5f);
     }
 
     public void Deinitialize()
@@ -181,17 +184,16 @@ public class HardcoreModeHandler
             Client.HHero.SetLives(0);
         
         // Check Dead Run
-        if (HardcoreRunDead && Client.HHero.GetLives() >= 1)
+        if (HardcoreRunDead && Client.HHero.GetLives() > 0)
         {
             Client.HHero.KillPlayer();
+            Client.HHero.SetHealth(1);
+            Client.HHero.SetWaterHealth(1);
             return;
         }
 
         var heroState = Client.HHero.GetHeroState();
         var isBull = Client.HHero.IsBull();
-        // Check Flinch
-        if (!isBull && heroState == (int)HeroState.KnockedOver)
-            Client.HHero.KillPlayer();
         // Read Breath
         var breath = Client.HHero.GetWaterHealth();
         // Set Breath
@@ -203,16 +205,174 @@ public class HardcoreModeHandler
         if (health > 1)
             Client.HHero.SetHealth(1);
         // Check Dead
-        if (health < 1 || breath < 1)
+        if (health < 1 || breath < 1 || (!isBull && heroState == (int)HeroState.KnockedOver))
         {
-            if (HardcoreRunDead)
-                return;
-            Client._client.Send(Message.Create(MessageSendMode.Reliable, MessageID.HC_RunStatusChanged).AddBool(true));
-            HardcoreRunDead = true;
+            EndRun();
             return;
         }
 
-        var count = SyncHandler.HThEg.GlobalObjectData.Values.Take(3).Sum(array => array.Count(x => x == (byte)5));
+        CheckThEggCounts();
+        
+        // EMU TIMER
+        ProcessHandler.TryRead(0x284AA8, out float emuTimer, true, "emuTimer");
+        if (emuTimer > 5000f)
+            ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x284AA8, BitConverter.GetBytes(5000f));
+        
+        // TURKEY AND BOONIE SPIN
+        ProcessHandler.TryRead(0x25D424, out int turkeyCount, true, "getTurkey");
+        if (turkeyCount != 0)
+        {
+            ProcessHandler.TryRead(0x25D428, out int turkeyAddr, true, "getTurkey");
+            for (var turkeyIndex = 0; turkeyIndex < turkeyCount; turkeyIndex++)
+            {
+                var rotationSpeed = _random.Next(0, 2) == 0 ? -0.3f : 0.3f;
+                ProcessHandler.WriteData(turkeyAddr + 0x2E8 + turkeyIndex * 0x4A0, BitConverter.GetBytes(rotationSpeed));
+            }
+        }
+        ProcessHandler.TryRead(0x257D5C, out int badBoonieCount, true, "getBadBoonie");
+        if (badBoonieCount != 0)
+        {
+            ProcessHandler.TryRead(0x257D60, out int badBoonieAddr, true, "getBadBoonie");
+            for (var badBoonieIndex = 0; badBoonieIndex < badBoonieCount; badBoonieIndex++)
+            {
+                var rotationSpeed = _random.Next(0, 2) == 0 ? -0.3f : 0.3f;
+                ProcessHandler.WriteData(badBoonieAddr + 0x2E8 + badBoonieIndex * 0x4A0, BitConverter.GetBytes(rotationSpeed));
+            }
+        }
+
+        const float arsonFrillSpawnDelay = 2.5f;
+        ProcessHandler.TryRead(0x289D2C, out int arsonFrillSpawnTimer, true, "spawnTimerArson");
+        if (arsonFrillSpawnTimer > (int)(arsonFrillSpawnDelay * 60))
+            ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x289D2C, BitConverter.GetBytes((int)(arsonFrillSpawnDelay * 60)));
+        
+        // Increase Rex Speed
+        var rexAddr = PointerCalculations.GetPointerAddress(0x25FF6C, new[] { 0x9C, 0x114 });
+        ProcessHandler.TryRead(rexAddr, out float rexSpeed, false, "rexSpeed");
+        if (Math.Abs(rexSpeed - 21.125f) < 0.1)
+            ProcessHandler.WriteData(rexAddr, BitConverter.GetBytes(21.5f));
+        if (Math.Abs(rexSpeed - 9.75) < 0.1)
+            ProcessHandler.WriteData(rexAddr, BitConverter.GetBytes(18f));
+        if (Math.Abs(rexSpeed - 6.5f) < 0.1)
+            ProcessHandler.WriteData(rexAddr, BitConverter.GetBytes(12.5f));
+
+
+        ProcessHandler.TryRead(0x2704FC, out int taCount, true, "taCount");
+        if (taCount != 0)
+        {
+            var inTimeAttack = Client.HGameState.IsInTimeAttack();
+            Client.HGameState.SetMenuItemFlag(TyMenuItem.ExitLevel, TyMenuItemFlag.Enabled, !inTimeAttack);
+            Client.HGameState.SetMenuItemFlag(TyMenuItem.MainMenu, TyMenuItemFlag.Enabled, !inTimeAttack);
+            ProcessHandler.TryRead(0x270500, out int taAddr, true, "taAddr");
+            ProcessHandler.TryRead(taAddr + 0xB4, out int loseDialogAddr, false, "loseDialogAddr");
+            ProcessHandler.TryRead(0x28C318, out int lastLoadedDialogAddr, true, "lastDialogAddr");
+            if (lastLoadedDialogAddr == loseDialogAddr)
+            {
+                EndRun();
+                return;
+            }
+        }
+        
+        // PERIMETER CHECK 
+        var addr = PointerCalculations.GetPointerAddress(0x2656F0, new[] { 0x6C });
+        ProcessHandler.TryRead(addr, out bool active, false, "Check Perimeter Check");
+        if (active && Client.HObjective.GetPerimeterCheckHealth() > 2)
+            Client.HObjective.SetPerimeterCheckHealth(2);
+        
+        // Handle Icicles
+        ProcessHandler.TryRead(0x26B964, out int icicleCount, true, "icicleCount");
+        if (icicleCount != 0)
+        {
+            ProcessHandler.TryRead(0x26B968, out int icicleAddr, true, "icicleAddress");
+            HDC_IcicleHandler.HandleIcicles(icicleCount, icicleAddr, (HDC_IcicleBehaviour)_random.Next(0, Enum.GetValues(typeof(HDC_IcicleBehaviour)).Length));
+        }
+    }
+    
+    public void SetEnemySpeedMultiplier(float multiplier = 1)
+    {
+        //FRILL
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25AB18, BitConverter.GetBytes(5f * multiplier));
+        //RUFUS
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25D0F8, BitConverter.GetBytes(7.0f * multiplier));
+        //CRAB
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25BFD0, BitConverter.GetBytes(5.33f * multiplier));
+        //MUDCRAB
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25BDF0, BitConverter.GetBytes(6.0f * multiplier));
+        //BLUETONGUE
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x2584AC, BitConverter.GetBytes(1.2f * multiplier));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x2584B0, BitConverter.GetBytes(4.3f * multiplier));
+        //SKINK
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25E000, BitConverter.GetBytes(4.3f * multiplier));
+        //SHARK
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25DD60, BitConverter.GetBytes(6.666f * multiplier));
+        //BIKERFRILL
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25A898, BitConverter.GetBytes(float.Min(32.50f * multiplier, 50f)));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25A89C, BitConverter.GetBytes(0.0192f * (float)Math.Pow(multiplier, 1)));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25A8A0, BitConverter.GetBytes(0.0997475f * (float)Math.Pow(multiplier, 1)));
+        //TURKEY
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25D470, BitConverter.GetBytes(11.50f * (float)Math.Pow(multiplier, 0.5f)));
+        //BOONIE
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x257DA8, BitConverter.GetBytes(11.50f * multiplier));
+        //WOMBAT
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25F6E8, BitConverter.GetBytes(10.50f * multiplier));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25F6EC, BitConverter.GetBytes(0.02166f * (float)Math.Pow(multiplier, 2)));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25F6F0, BitConverter.GetBytes(0.045f * (float)Math.Pow(multiplier, 2)));
+        //ANDY
+        switch (multiplier)
+        {
+            case 1:
+                ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25FA28, BitConverter.GetBytes(5f * multiplier));
+                ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25FA30, BitConverter.GetBytes(0.045f * multiplier));
+                break;
+            case 2:
+                ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25FA28, BitConverter.GetBytes(8.15f * multiplier));
+                ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25FA30, BitConverter.GetBytes(0.2f * multiplier));
+                break;
+        }
+
+        //ARSONFRILL
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25AE48, BitConverter.GetBytes(5.0f * multiplier));
+        //LILNEDDY
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25BB60, BitConverter.GetBytes(4.7f * multiplier));
+        //SLY
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25E438, BitConverter.GetBytes(4.0f * multiplier));
+        //BARRACUDA
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x258010, BitConverter.GetBytes(8.00f * multiplier));
+        //NEDDY
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C23C, BitConverter.GetBytes(1.60f * multiplier));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C240, BitConverter.GetBytes(20f * (float)Math.Pow(multiplier, 2)));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C244, BitConverter.GetBytes(0.02166f * (float)Math.Pow(multiplier, 2)));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C2A0, BitConverter.GetBytes(50f * (float)Math.Pow(multiplier, 2)));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C2A4, BitConverter.GetBytes(180f * (float)Math.Pow(multiplier, 2)));
+        //DENNIS
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x259A98, BitConverter.GetBytes(6.0f * multiplier));
+        //CHEMICALFRILL
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25ACAC, BitConverter.GetBytes(2.0f * multiplier));
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25ACB4, BitConverter.GetBytes(0.032724f * (float)Math.Pow(multiplier, 2.5f)));
+        //RHINO
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25CBBC, BitConverter.GetBytes(3.30f * multiplier));
+        //CROC
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x2597B8, BitConverter.GetBytes(5.50f * multiplier));
+        //NINJA
+        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C4F0, BitConverter.GetBytes(10.40f * multiplier));
+    }
+    
+    [MessageHandler((ushort)MessageID.HC_RunStatusChanged)]
+    public static void HandleRunStatusChanged(Message message)
+    {
+        Client.HHardcore.HardcoreRunDead = message.GetBool();
+    }
+
+    public void EndRun()
+    {
+        if (HardcoreRunDead)
+            return;
+        Client._client.Send(Message.Create(MessageSendMode.Reliable, MessageID.HC_RunStatusChanged).AddBool(true));
+        HardcoreRunDead = true;
+    }
+
+    private void CheckThEggCounts()
+    {
+         var count = SyncHandler.HThEg.GlobalObjectData.Values.Take(3).Sum(array => array.Count(x => x == (byte)5));
         if (count >= 17 && SyncHandler.HAttribute.GlobalObjectData[(int)Attributes.GotFlame] == 0)
         {
             var address = SyncHandler.SaveDataBaseAddress + 0xA84;
@@ -247,99 +407,5 @@ public class HardcoreModeHandler
             ProcessHandler.WriteData(address + (int)Attributes.GotTali3, new byte[] { 1 });
             SFXPlayer.PlaySound(SFX.RangGet);
         }
-        
-        ProcessHandler.TryRead(0x25D424, out int turkeyCount, true, "getTurkey");
-        if (turkeyCount != 0)
-        {
-            ProcessHandler.TryRead(0x25D428, out int turkeyAddr, true, "getTurkey");
-            for (var turkeyIndex = 0; turkeyIndex < turkeyCount; turkeyIndex++)
-            {
-                var rotationSpeed = _random.Next(0, 2) == 0 ? -0.3f : 0.3f;
-                ProcessHandler.WriteData(turkeyAddr + 0x2E8 + turkeyIndex * 0x4A0, BitConverter.GetBytes(rotationSpeed));
-            }
-        }
-
-        const float arsonFrillSpawnDelay = 2.5f;
-        ProcessHandler.TryRead(0x289D2C, out int arsonFrillSpawnTimer, true, "spawnTimerArson");
-        if (arsonFrillSpawnTimer > (int)(arsonFrillSpawnDelay * 60))
-            ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x289D2C, BitConverter.GetBytes((int)(arsonFrillSpawnDelay * 60)));
-        
-        // Increase Rex Speed
-        var rexAddr = PointerCalculations.GetPointerAddress(0x25FF6C, new[] { 0x9C, 0x114 });
-        ProcessHandler.TryRead(rexAddr, out float rexSpeed, false, "rexSpeed");
-        if (Math.Abs(rexSpeed - 21.125f) < 0.1)
-            ProcessHandler.WriteData(rexAddr, BitConverter.GetBytes(22f));
-        if (Math.Abs(rexSpeed - 9.75) < 0.1)
-            ProcessHandler.WriteData(rexAddr, BitConverter.GetBytes(22f));
-        if (Math.Abs(rexSpeed - 6.5f) < 0.1)
-            ProcessHandler.WriteData(rexAddr, BitConverter.GetBytes(12.5f));
-        
-        // PERIMETER CHECK 
-        var addr = PointerCalculations.GetPointerAddress(0x2656F0, new[] { 0x6C });
-        ProcessHandler.TryRead(addr, out bool active, false, "Check Perimeter Check");
-        if (active && Client.HObjective.GetPerimeterCheckHealth() > 2)
-            Client.HObjective.SetPerimeterCheckHealth(2);
-    }
-    
-    public void SetEnemySpeedMultiplier(float multiplier = 1)
-    {
-        //FRILL
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25AB18, BitConverter.GetBytes(5f * multiplier));
-        //RUFUS
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25D0F8, BitConverter.GetBytes(7.0f * multiplier));
-        //CRAB
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25BFD0, BitConverter.GetBytes(5.33f * multiplier));
-        //MUDCRAB
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25BDF0, BitConverter.GetBytes(6.0f * multiplier));
-        //BLUETONGUE
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x2584AC, BitConverter.GetBytes(1.2f * multiplier));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x2584B0, BitConverter.GetBytes(4.3f * multiplier));
-        //SKINK
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25E000, BitConverter.GetBytes(4.3f * multiplier));
-        //SHARK
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25DD60, BitConverter.GetBytes(6.666f * multiplier));
-        //BIKERFRILL
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25A898, BitConverter.GetBytes(float.Min(32.50f * multiplier, 50f)));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25A89C, BitConverter.GetBytes(0.0192f * (float)Math.Pow(multiplier, 1)));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25A8A0, BitConverter.GetBytes(0.0997475f * (float)Math.Pow(multiplier, 1)));
-        //TURKEY
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25D470, BitConverter.GetBytes(11.50f * (float)Math.Pow(multiplier, 0.5f)));
-        //BOONIE
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x257DA8, BitConverter.GetBytes(11.50f * multiplier));
-        //WOMBAT
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25F6E8, BitConverter.GetBytes(10.50f * multiplier));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25F6EC, BitConverter.GetBytes(0.02166f * (float)Math.Pow(multiplier, 2)));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25F6F0, BitConverter.GetBytes(0.045f * (float)Math.Pow(multiplier, 2)));
-        //ARSONFRILL
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25AE48, BitConverter.GetBytes(5.0f * multiplier));
-        //LILNEDDY
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25BB60, BitConverter.GetBytes(4.7f * multiplier));
-        //SLY
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25E438, BitConverter.GetBytes(4.0f * multiplier));
-        //BARRACUDA
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x258010, BitConverter.GetBytes(8.00f * multiplier));
-        //NEDDY
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C23C, BitConverter.GetBytes(1.60f * multiplier));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C240, BitConverter.GetBytes(20f * (float)Math.Pow(multiplier, 2)));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C244, BitConverter.GetBytes(0.02166f * (float)Math.Pow(multiplier, 2)));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C2A0, BitConverter.GetBytes(50f * (float)Math.Pow(multiplier, 2)));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C2A4, BitConverter.GetBytes(180f * (float)Math.Pow(multiplier, 2)));
-        //DENNIS
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x259A98, BitConverter.GetBytes(6.0f * multiplier));
-        //CHEMICALFRILL
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25ACAC, BitConverter.GetBytes(2.0f * multiplier));
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25ACB4, BitConverter.GetBytes(0.032724f * (float)Math.Pow(multiplier, 2.5f)));
-        //RHINO
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25CBBC, BitConverter.GetBytes(3.30f * multiplier));
-        //CROC
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x2597B8, BitConverter.GetBytes(5.50f * multiplier));
-        //NINJA
-        ProcessHandler.WriteData((int)TyProcess.BaseAddress + 0x25C4F0, BitConverter.GetBytes(10.40f * multiplier));
-    }
-    
-    [MessageHandler((ushort)MessageID.HC_RunStatusChanged)]
-    public static void HandleRunStatusChanged(Message message)
-    {
-        Client.HHardcore.HardcoreRunDead = message.GetBool();
     }
 }