+++ Mul-Ty-Player Updater/SettingsHandler.cs
@@ -35,11 +35,5 @@ public class Settings
     public string ClientDir { get; set; }
     public string ServerDir { get; set; }
     public string GameDir { get; set; }
-    public bool RevertOutbackMovement { get; set; }
-    public bool RevertRangSwitching { get; set; }
-    public bool FixControllerCameraAiming { get; set; }
-    public bool OpenAllGameInfo { get; set; }
-    public bool FixMenuBug { get; set; }
-    public bool FixedMagnets { get; set; }
     public string Version { get; set; }
 }